Add HashtagGenerator and TextPreprocessing classes
Added HashtagGenerator.java and TextPreprocessing.java to src/main/java/org/epam/engx. The HashtagGenerator class is created to generate hashtags from a given text using natural language processing techniques. It uses the OpenNLP library for tokenization and part-of-speech tagging, hence it is necessary to update dependencies in pom.xml file. The TextPreprocessing class is a helper class used in the pre-processing step to transform the input text to a format suitable for hashtag generation. It removes non-alphabetic characters and converts text to lowercase.